Fucking fixed fucking wew !!!
Turned out the problem was due to two things:

1. XCB_MASK_EVENT_KEY_PRESS
2. switch (e->response_type & ~0x80 & mask)

The KEY_PRESS event mask can be registered by ONLY ONE application.
Which is why, when wew was registered to it, other applications trying
to access it would fail with BadAccess error.
This event was registered because -otherwise- ENTER_WINDOW event would
not trigger (that's whas I though)

Turns out it was due to the "& mask" in the switch statement, forbidding
events to be handled if they were not in the mask set up by wew. The
problem here is that the mask used to register event is TOTALLY
different from to values of the event triggered, so

> XCB_EVENT_MASK_ENTER_WINDOW == 14
> XCB_ENTER_NOTIFY == 7

But adding together random masks (including KEY_PRESS) made the
ENTER_NOTIFY event be raised.

Damn I'm so stupid.