Removed xmmv(1) from the opt/ repo
wmv/wrs can now be used with absolute positions, making useful to
move/resize windows using the mouse.

For example, to "extend" the window to the position of the cursor, you
can do

	wrs -a $(wmp) $(pfw)

To move the window to the position of the cursor

	wmv -a $(wmp) $(pfw)

You get it. For instance, sxhkd(1) has a nice interface that makes it
possible to mv/rs windows using the mouse:

	alt + !button{1,3}
		{wmv,wrs} -a %i %i $(pfw)

This will let you use wmv/wrs with the mouse. Neat.