Revert "wew: Fix enter/leave events"
This reverts commit 8754fc9ddcf31d99f79e214088389eb0af3e2966.

The fix introduced brings way to much problems in the daily workflow.
The window under the cursor always has the focus, resulting in moving
windows awkwardly, hiding mouse menus behind the primary window and
other similar issues.

It forces the dev to move the cursor along with the windows, resulting
in twice as much code, for no particular reason.