wname: Let atom WM_NAME be of any type
Traditionnaly, the atom WM_NAME is of type XCB_ATOM_STRING. However,
some recent application will create a new type (eg. UTF8_STRING) and
use that instead, which means that we won't be able to retrieve it when
limiting ourself to a single type.

Accepting atom to be of any type, then casting it to a (char *), while
not perfect, will cover most use cases.